feat(education): the seed importer, reusing the contract verbatim
seed_import imports scripts/education_seed_contract.py itself -- the same
module validate_education_seed runs in CI -- so the importer cannot hold a
looser idea of 'valid' than the gate does. A test feeds it a bundle the
contract rejects and asserts nothing was written: validation runs against
the whole bundle before the first session.add, because a partial import
leaves a corpus that looks populated and is wrong.

Two-pass on institutions: parent_slug and merged_into_slug can point at a
row further down the file, so the self-references wait until every
institution has an id.

Three things the plan did not account for, all found by running it:

- Every test needed owner_session, not db_session. The SELECT-only grant
  means an app_rt fixture cannot seed a college. Same split as the model
  suite.
- The test database has no geo. A new education_geo fixture loads the real
  36-state snapshot first, because state_id is an FK and an empty geo table
  means either a constraint error or 772 rows with a null state -- and the
  second one looks healthy until every state page comes up empty. The
  importer now refuses outright on empty geo, with a test.
- lat/lng were annotated float against a Numeric column. asyncpg
  round-trips Numeric as Decimal, and shared/geo/models.py declares its
  centroids that way. Corrected to Decimal in the model and the importer.

Also renamed the reused loop variable, which was named "obj" in all five
table loops. mypy reported 88 errors from it: one variable across five
loops pins its type to the first, so every later assignment looked like an
error on Institution.

+++ backend/core/modules/education/models.py
@@ -13,6 +13,7 @@ from __future__ import annotations
 
 import uuid
 from datetime import date
+from decimal import Decimal
 from typing import Any
 
 from sqlalchemy import Boolean, Date, ForeignKey, Integer, Numeric, Text, UniqueConstraint
@@ -59,8 +60,12 @@ class Institution(UUIDv7PKMixin, ImmutableSlugMixin, TimestampMixin, Base):
     )
     district_id: Mapped[int | None] = mapped_column(Integer, index=True)
     pincode: Mapped[str | None] = mapped_column(Text)
-    lat: Mapped[float | None] = mapped_column(Numeric(9, 6))
-    lng: Mapped[float | None] = mapped_column(Numeric(9, 6))
+    # Decimal, not float: Numeric round-trips as Decimal through asyncpg, and
+    # shared/geo/models.py declares its centroids the same way. A float
+    # annotation would be a lie the type checker happens not to catch until
+    # something assigns to it.
+    lat: Mapped[Decimal | None] = mapped_column(Numeric(9, 6))
+    lng: Mapped[Decimal | None] = mapped_column(Numeric(9, 6))
     address: Mapped[str | None] = mapped_column(Text)
     website: Mapped[str | None] = mapped_column(Text)
     contact_phone: Mapped[str | None] = mapped_column(Text)
